# Conflicts:
#	app/src/main/kotlin/dev/aaa1115910/bv/viewmodel/home/RecommendViewModel.kt
#	gradle/androidx.versions.toml

+++ app/src/main/kotlin/dev/aaa1115910/bv/repository/VideoInfoRepository.kt
@@ -5,8 +5,8 @@ class VideoInfoRepository {
 }
 
 data class VideoListItem(
-    val aid: Int,
-    val cid: Int,
+    val aid: Long,
+    val cid: Long,
     val epid: Int? = null,
     val seasonId: Int? = null,
     val title: String,
